Move Canvas into //flutter/lib/ui
This patch moves the remainder of painting and compositing into their
new home (and updates their style to match Google C++ Style).
