+++ flutter/lib/ui/painting/picture.cc
@@ -4,11 +4,11 @@
 
 #include "flutter/lib/ui/painting/picture.h"
 
+#include "flutter/lib/ui/painting/canvas.h"
 #include "flutter/tonic/dart_args.h"
 #include "flutter/tonic/dart_binding_macros.h"
 #include "flutter/tonic/dart_converter.h"
 #include "flutter/tonic/dart_library_natives.h"
-#include "sky/engine/core/painting/Canvas.h"
 
 namespace blink {
