+++ sky/engine/bindings/dart_ui.cc
@@ -4,22 +4,22 @@
 
 #include "sky/engine/bindings/dart_ui.h"
 
+#include "flutter/lib/ui/compositing/scene.h"
+#include "flutter/lib/ui/compositing/scene_builder.h"
+#include "flutter/lib/ui/painting/canvas.h"
 #include "flutter/lib/ui/painting/gradient.h"
 #include "flutter/lib/ui/painting/image.h"
+#include "flutter/lib/ui/painting/image_decoding.h"
 #include "flutter/lib/ui/painting/image_filter.h"
 #include "flutter/lib/ui/painting/image_shader.h"
 #include "flutter/lib/ui/painting/mask_filter.h"
 #include "flutter/lib/ui/painting/path.h"
 #include "flutter/lib/ui/painting/picture.h"
+#include "flutter/lib/ui/painting/picture_recorder.h"
 #include "flutter/tonic/dart_converter.h"
 #include "flutter/tonic/dart_error.h"
 #include "sky/engine/bindings/dart_runtime_hooks.h"
 #include "sky/engine/bindings/mojo_services.h"
-#include "sky/engine/core/compositing/Scene.h"
-#include "sky/engine/core/compositing/SceneBuilder.h"
-#include "sky/engine/core/painting/Canvas.h"
-#include "sky/engine/core/painting/PictureRecorder.h"
-#include "sky/engine/core/painting/painting.h"
 #include "sky/engine/core/text/Paragraph.h"
 #include "sky/engine/core/text/ParagraphBuilder.h"
 #include "sky/engine/core/window/window.h"
@@ -49,11 +49,11 @@ void DartUI::InitForGlobal() {
     CanvasImage::RegisterNatives(g_natives);
     CanvasPath::RegisterNatives(g_natives);
     DartRuntimeHooks::RegisterNatives(g_natives);
+    ImageDecoding::RegisterNatives(g_natives);
     ImageFilter::RegisterNatives(g_natives);
     ImageShader::RegisterNatives(g_natives);
     MaskFilter::RegisterNatives(g_natives);
     MojoServices::RegisterNatives(g_natives);
-    Painting::RegisterNatives(g_natives);
     Paragraph::RegisterNatives(g_natives);
     ParagraphBuilder::RegisterNatives(g_natives);
     Picture::RegisterNatives(g_natives);
