+++ sky/engine/core/core.gni
@@ -9,21 +9,11 @@ sky_core_files = [
   "Init.h",
   "start_up.h",
   "start_up.cc",
-  "compositing/Scene.cpp",
-  "compositing/Scene.h",
-  "compositing/SceneBuilder.cpp",
-  "compositing/SceneBuilder.h",
   "editing/CompositionUnderline.h",
   "editing/CompositionUnderlineRangeFilter.cpp",
   "editing/CompositionUnderlineRangeFilter.h",
   "editing/PositionWithAffinity.cpp",
   "editing/PositionWithAffinity.h",
-  "painting/Canvas.cpp",
-  "painting/Canvas.h",
-  "painting/painting.cc",
-  "painting/painting.h",
-  "painting/PictureRecorder.cpp",
-  "painting/PictureRecorder.h",
   "rendering/BidiRun.h",
   "rendering/BidiRunForLine.cpp",
   "rendering/BidiRunForLine.h",
@@ -189,7 +179,6 @@ if (!flutter_product_mode) {
 }
 
 core_dart_files = get_path_info([
-                                  "dart/compositing.dart",
                                   "dart/hash_codes.dart",
                                   "dart/hooks.dart",
                                   "dart/lerp.dart",
