+++ sky/engine/core/dart/text.dart
@@ -681,6 +681,11 @@ abstract class Paragraph extends NativeFieldWrapperClass2 {
   /// Word boundaries are defined more precisely in Unicode Standard Annex #29
   /// http://www.unicode.org/reports/tr29/#Word_Boundaries
   List<int> getWordBoundary(int offset) native "Paragraph_getWordBoundary";
+
+  // Redirecting the paint function in this way solves some dependency problems
+  // in the C++ code. If we straighten out the C++ dependencies, we can remove
+  // this indirection.
+  void _paint(Canvas canvas, double x, double y) native "Paragraph_paint";
 }
 
 /// Builds a [Paragraph] containing text with the given styling information.
